transferindo o método draw e getScale da classe Slide para SlideViewerComponent
