+++ src/main/java/br/upe/ppsw/jabberpoint/apresentacao/XMLAccessor.java
@@ -149,7 +149,7 @@ public class XMLAccessor extends Accessor {
 			Slide slide = presentation.getSlide(slideNumber);
 
 			out.println("<slide>");
-			out.println("<title>" + slide.getTitle() + "</title>");
+			out.println("<title>" + slide.getTitleText() + "</title>");
 
 			Vector<SlideItem> slideItems = slide.getSlideItems();
 			for (int itemNumber = 0; itemNumber < slideItems.size(); itemNumber++) {
